refactor: refactor file contents
Uses normal package naming and conventions. Adds tests for the
RandomBobifier.
